refactor workbench performance module routes
