+++ src/app/router_registry.py
@@ -64,6 +64,9 @@ from app.routers.workbench_performance_advisor_brief import (
 from app.routers.workbench_performance_evidence import (
     router as workbench_performance_evidence_router,
 )
+from app.routers.workbench_performance_modules import (
+    router as workbench_performance_modules_router,
+)
 from app.routers.workbench_risk import router as workbench_risk_router
 from app.routers.workbench_sandbox import router as workbench_sandbox_router
 
@@ -106,6 +109,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(workbench_performance_router)
     app.include_router(workbench_performance_advisor_brief_router)
     app.include_router(workbench_performance_evidence_router)
+    app.include_router(workbench_performance_modules_router)
     app.include_router(workbench_risk_router)
     app.include_router(workbench_sandbox_router)
     app.include_router(reporting_router)
